chore: Update event properties and types;
