+++ apps/web/modules/calendar/components/week-and-day-view/day-view-multi-day-events-row.tsx
@@ -2,7 +2,7 @@ import { differenceInDays, endOfDay, isWithinInterval, parseISO, startOfDay } fr
 
 import { MonthEventBadge } from "@/modules/calendar/components/month-view/month-event-badge";
 
-import type { IEvent } from "@/modules/calendar/interfaces";
+import type { IEvent } from "@/modules/calendar/types";
 
 interface IProps {
   selectedDate: Date;
